+++ pywalletconnect/enc_tunnel.py
@@ -20,16 +20,36 @@
 from json import loads
 from os import urandom
 
-from cryptography.hazmat.primitives import hashes, hmac
+from cryptography.hazmat.primitives import hashes, hmac, serialization
 from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
 from cryptography.hazmat.primitives.padding import PKCS7
-
+from cryptography.hazmat.primitives.asymmetric.x25519 import (
+    X25519PrivateKey,
+    X25519PublicKey,
+)
 
 # ---- Crypto primitives
 
 
 AES_BLOCK_SIZE = 16
 AES_BLOCK_SIZE_BITS = AES_BLOCK_SIZE << 3
+
+
+def msg_digest(msg, halgo):
+    """Base secure hash standard function."""
+    digest = hashes.Hash(halgo)
+    digest.update(msg)
+    return digest.finalize()
+
+
+def sha2_256(msg):
+    """Compute SHA256."""
+    return msg_digest(msg, hashes.SHA256())
+
+
+def sha2_512(msg):
+    """Compute SHA512."""
+    return msg_digest(msg, hashes.SHA512())
 
 
 def hmac_sha256(key, message):
@@ -83,7 +103,7 @@ def decrypt_aes(key, init_vector, message):
 
 
 class EncryptedTunnel:
-    """Provide an encryption tunnel for WalletConnect messages."""
+    """Provide an encryption tunnel for WalletConnect v1 messages."""
 
     def __init__(self, key):
         """Start a tunnel with an AES key."""
@@ -108,3 +128,72 @@ class EncryptedTunnel:
         init_vector = bytes.fromhex(payload_obj["iv"])
         check_hmac(msg_bin + init_vector, mac_sig, self.key)
         return decrypt_aes(self.key, init_vector, msg_bin)
+
+
+class EncryptedTunnelv2:
+    """Provide an encryption tunnel for WalletConnect v2 messages."""
+
+    def __init__(self, pubkey, enc_key, mac_key):
+        """Start a v2 tunnel."""
+        self.enckey = enc_key
+        self.mackey = mac_key
+        self.pubkey = pubkey
+
+    def encrypt_payload(self, message_str, encod="utf8"):
+        """Encrypt a string or bytes message into a hex message.
+        If encod is None or "", the message is read as binary bytes.
+        iv, publicKey, mac and cipherText
+        """
+        if encod:
+            message = message_str.encode(encod)
+        else:
+            message = message_str
+        enc_msg_iv = encrypt_aes(self.enckey, message)
+        mac_data = hmac_sha256(
+            self.mackey, enc_msg_iv[1] + self.pubkey + enc_msg_iv[0]
+        ).hex()
+        return enc_msg_iv[1].hex() + self.pubkey.hex() + mac_data + enc_msg_iv[0].hex()
+
+    def decrypt_payload(self, fullpayload_hex):
+        """Decrypt a payload object into a bytes message."""
+        init_vector = bytes.fromhex(fullpayload_hex[:32])
+        pubkey = bytes.fromhex(fullpayload_hex[32:96])
+        mac_sig = bytes.fromhex(fullpayload_hex[96:160])
+        enc_msg = bytes.fromhex(fullpayload_hex[160:])
+        check_hmac(init_vector + pubkey + enc_msg, mac_sig, self.mackey)
+        return decrypt_aes(self.enckey, init_vector, enc_msg).decode("utf8")
+
+
+# ---- Asymmetric key pair class for v2
+
+
+class KeyAgreement:
+    """X25519 key pair."""
+
+    def __init__(self):
+        """Generate an X25519 key pair."""
+        self.key = X25519PrivateKey.generate()
+        self.shared_key = None
+
+    def get_pubkey(self):
+        """Give the local public key."""
+        return self.key.public_key().public_bytes(
+            encoding=serialization.Encoding.Raw, format=serialization.PublicFormat.Raw
+        )
+
+    def compute_shared_key(self, peer_key_hex):
+        """Compute shared key from the peer proposer public key."""
+        peer_public_key = X25519PublicKey.from_public_bytes(bytes.fromhex(peer_key_hex))
+        self.shared_key = self.key.exchange(peer_public_key)
+
+    def derive_enc_key(self):
+        """uses a SHA512 hash of the shared key
+        using the first 32bytes for encryption
+        and the last 32 bytes for authentication.
+        """
+        hashk = sha2_512(self.shared_key)
+        return hashk[:32], hashk[32:]
+
+    def derive_topic(self):
+        """Topic for the next sequence."""
+        return sha2_256(self.shared_key).hex()